feat(ui): show template diff in update modal
