+++ ui/src/app/shared/workflow-template/modal/workflow-template.modal.html
@@ -16,7 +16,7 @@
                     </div>
                 </div>
                 <div class="sixteen wide column" *ngIf="workflowTemplate && workflowTemplateInstance">
-                    <div class="ui wide eighteen field">
+                    <div class="ui wide sixteen field">
                         <div class="fields">
                             <div class="six wide field">
                                 <label>{{'common_name' | translate}}</label>
@@ -27,14 +27,17 @@
                                 <input class="ui input" type="text" name="group" [disabled]="true" [ngModel]="workflowTemplate.group.name">
                             </div>
                             <div class="four wide field right aligned">
-                                <a *ngIf="workflowTemplateInstance.workflow_template_version === workflowTemplate.version"
-                                    class="ui green label">{{'common_up_to_date'
-                                    | translate}}</a>
+                                <div *ngIf="workflowTemplateInstance.workflow_template_version === workflowTemplate.version"
+                                    class="ui label green"> {{'common_up_to_date' | translate}}</div>
                                 <a *ngIf="workflowTemplateInstance.workflow_template_version !== workflowTemplate.version"
-                                    class="ui orange label">{{'common_not_up_to_date' | translate }}</a>
+                                    class="ui orange label" (click)="toggleDiff()">{{'common_not_up_to_date' |
+                                    translate
+                                    }}</a>
                             </div>
                         </div>
-                        <app-diff-list *ngIf="diffItems" [items]="diffItems"></app-diff-list>
+                        <div class="ui wide sixteen field">
+                            <app-diff-list *ngIf="diffVisible" [items]="diffItems"></app-diff-list>
+                        </div>
                         <app-workflow-template-apply-form [project]="project" [workflowTemplate]="workflowTemplate"
                             [workflowTemplateInstance]="workflowTemplateInstance" [withClose]="true" (close)="close()"
                             (apply)="loadInstance()"></app-workflow-template-apply-form>
